fix: shake some images from server bundle
