+++ packages/client/components/ActivityLibrary/Categories.ts
@@ -1,11 +1,4 @@
 import {MeetingTypeEnum} from '../../__generated__/NewMeetingQuery.graphql'
-import retroBackgroundSrc from '../../../../static/images/illustrations/retro-background.png'
-import standupBackgroundSrc from '../../../../static/images/illustrations/standup-background.png'
-import feedbackBackgroundSrc from '../../../../static/images/illustrations/feedback-background.png'
-import estimationBackgroundSrc from '../../../../static/images/illustrations/estimation-background.png'
-import strategyBackgroundSrc from '../../../../static/images/illustrations/strategy-background.png'
-import premortemBackgroundSrc from '../../../../static/images/illustrations/premortem-background.png'
-import postmortemBackgroundSrc from '../../../../static/images/illustrations/postmortem-background.png'
 import {CardTheme} from './ActivityCard'
 
 export const MAIN_CATEGORIES = [
@@ -54,13 +47,3 @@ export const MEETING_TYPE_TO_CATEGORY: Record<MeetingTypeEnum, CategoryID> = {
   poker: 'estimation',
   teamPrompt: 'standup'
 }
-
-export const backgroundImgMap = {
-  retrospective: retroBackgroundSrc,
-  standup: standupBackgroundSrc,
-  feedback: feedbackBackgroundSrc,
-  estimation: estimationBackgroundSrc,
-  strategy: strategyBackgroundSrc,
-  premortem: premortemBackgroundSrc,
-  postmortem: postmortemBackgroundSrc
-} as const
